+++ src/PageContainer.js
@@ -6,7 +6,7 @@ import PageHeader from './PageLayout/PageHeader.js';
 import logo from './icons/temp-logo.svg';
 import RecurringPickupRequest from './PageContent/RequestPickup/RecurringPickupRequest.js';
 import AssignVolunteersController from './PageContent/AssignVolunteers/AssignVolunteersController.js';
-import Calendar from './PageContent/Calendar/Calendar';
+import Calendar from './PageContent/Calendar/ReactCalendar/Calendar';
 // The page to load when user is signed in.
 // Consist of the base page layout and page content depending on which tab is chosen.
 // Default page content is Calendar.
