+++ src/PageContent/Calendar/ReactCalendar/shared/locales.js
